mlxm-qwen5bmodel/requirements.txt
fastapi
uvicorn[standard]
mlx
mlx_lm
pyinstaller
pywebview 
streamlit 
requests
markdown